Fix instruction for building from source
Fixes #5

We cannnot build from source with `go get github.com/minamijoyo/tfupdate`
Because the hcl2 branch in hashicorp/hcl has not been merged into master yet.
I'm not sure this is an issue or a spec of `go get`,
but we need to fix the instruction in README anyway.