Preserve sort order and comments inside object for required_providers syntax
Updating the whole object loses original sort order and comments.
At the time of writing, there is no way to update a value inside an
object directly while preserving original tokens.

Since we fully understand the valid syntax, we compromise and read the
tokens in order, updating the bytes directly.
It's apparently a fragile dirty hack, but I didn't come up with the better
way to do this.