Sort releases in semver order
Fixes #36

Apparently I misunderstood the meaning of the latest release in GitHub.
The latest release just means the most recent release, not the latest
stable release.

We should not use the GetLatestRelease API in GitHub. We need to get all
releases, sort them in the semver order and find the latest stable
release.

I also investigated and found that it also seems to be affect in GitLab.
It also get releases order by released_at (default) or created_at, but
not semver, and there is no API or UI to set the latest release.

I'm not sure it also affects Terraform Registry becasuse I don't have
its account, but I think we should use the same strategy for
consistency.