Hide pre-releases by default in the release list command
Fixes #31

The release list command now hides pre-releases by default and allows it
to show by a new flag --pre-release.

This was originally requested in #31, but there was an implementation
concern for consistency at the time because all source types doesn't
have the same API metadata. In #41, the list is now sorted in semver
order and pre-releases can now be filtered in a consistent way for all
source types. Since the list command is originally for debug, it is
convenient to be able to switch the presence or absence of pre-releases.

There may be some opinions about whether the concept of release includes
pre-release or not, but it's clear that latest must not be pre-release,
and it seems convenient if the end of the list matches the latest for
debug. So change it to hide pre-releases by default.

This is a breaking change, but it's a good opportunity to change the
default in the next release because the order was changed by #41.