Move CI to GitHub Actions
The Docker Hub recently stopped auto builds for free plan.
We need to build and push docker images as a part of release process.

Because I'm maintaining multiple OSS projects,
using the same CI setup will reduce maintenance costs.
So I'll move the CI of this project to GitHub Actions.