Add examples for how to pin the AWS provider to v3
Closes #64

It's unclear that a value of version flag accepts any string literal in
README. That is, we can also pass a version constraint.
It's better to mention it explicitly in README.
I added examples for how to pin the AWS provider to v3.
I guess it's probably the next FAQ :wink: