Implement an index of provider versions for caching
Currently the Terraform Registry returns the zh hash for all platforms,
but not the h1 hash, so the h1 hash has to be calculated separately.
We need to calculate the values for each platform and merge the results.
To avoid multiple downloads and recalculations for each directory, the
results are cached in memory.