Implement version constraint detection
In order to update dependency lock files, we need to detect the version
constraints of the providers of the module, but the current Updater
interface can only handle a single file. Therefore, we introduce a new
concept called ModuleContext, which is supposed to provide the Updater
with information of the module that is currently being updated. Changes
to the Updater interface will be addressed later because it requires
massive rewriting.

Note that it does not actually re-implement the resolution of version
constraints in terraform init. It is very simplified for the use we
need. Version constraints only support simple constants and not
comparison operators. Ignore what cannot be interpreted.