Inject a ModuleContext to Updater
Until now, the Updater interface has been implemented statelessly, but
in order to adding a feature for updating the dependency lock file, we
also need the module context. I've only adjusted the interface for now,
but this will also help us to achieve more smart rewriting in the
future. We also need to access the index to cache the hash value of the
provider, which is a global state between modules. I decided to access
the global context via the module context because it is always needed in
combination with the module context. In addition, the option and updater
instances were also moved to the global context, as they have not
changed within the process.