Add native support for updating .terraform.lock.hcl
The Terraform dependency lock file (a.k.a. .terraform.lock.hcl) was
introduced in v0.14. While this is fine for simple use cases, there are
may challenges at scale. This is because terraform is based on
per-directory processing, and updating lock files using the official
terraform providers mirror and terraform providers lock commands is very
inefficient.

However, at that time, we avoid supporting for updating the lock file
for the following reasons:

- That the underlying reason for the complexity of updating the lock
  file is due to the lack of capability of the Terraform Registry; it
  does not return the h1 hash value.
- That the Registry protocol could be improved in the future.
- The lock file format implies to support not only providers but also
  modules.
- The lock file format is an implementation detail.
- It was easy to imagine that supporting multiple Terraform versions
  would be hard.

It got worse as time went on:

- With the rise of arm chips, the number of platforms required for
  locking has increased.
- As our infrastructure grown, the number of directories has increased
  accordingly.

However, no progress has been made on this issue in the upstream. So it
is time to reinvent the wheel by ourselves. Calculate hash values in
.terraform.lock.hcl and cache them to be able to update multiple
directories at once.