Fix a regression issue for using absolute path
Fixes #93

A ModuleContext was introduced to support the lock command in tfupdate
v0.7.0. This is intented to use the results of module inspection in any
updaters, which is currently not used except for the lock command, but
the inspection itself is enabled for all updaters. When implementing
this, there was a type miss match between afero and
terraform-config-inspect filesystems. We worked around by using a
compatibility layer via the standard io/fs.FS. However, it turned out
that afero.IOFS does not support absolute paths.
https://github.com/minamijoyo/tfupdate/issues/93

Ideally, it would be better to rewrite tests without afero, but it would
require a lot of work. To fix the regression issue as quickly as we can,
we will simply suppress the load error and add a validation for absolute
path on the lock command side because we actually don't need this
inspection result except for the lock command for now.