Fix a regression issue for updating .hcl file
Fixes #96

Since v0.7.0, to support .terraform.lock.hcl, I had to load the new .hcl
extension as well. Since the schema definitions of the provider block in
.terraform.lock.hcl and the provider block in .tf were different, the
filename was the only clue to distinguish the schemas. Therefore, while
expanding the range of filenames to be processed, I added a check of the
filename in each updater.

In v0.6.8, I thought that only .tf files were updated, but it was the
wrong assumption. v0.6.8 checked the file extension if the target was a
directory but did not check the file extension if the filename was
directly specified. In v0.7.0, each updater checks the filename and
ignores files other than .tf.

Strictly speaking, the .hcl extension is too generic to determine the
application schema. Therefore, processing directories is risky and
potentially collapses files of other applications. However, when you
specify the filename, not the directory, you intend to update it, so we
should accept it whenever possible. Fortunately or unfortunately, the
filename of the .terraform.lock.hcl file cannot be changed, so it is
safe to check this more strictly.

Even though this was an unintended use for me, I will treat it as a
regression and fix it.