Add RE2 regex support for module updates
Add new option `--source-match-type` to define the type of the match.
Possible values are `full` and `regex`. The default value is `full`.

The match type `full` is the same as the current behavior.

The match type `regex` is the new behavior. It uses RE2 regex to match
the module source URL with the given name pattern.

```bash
tfupdate module -v 2.14.1 --source-match-type regex terraform-aws-modules/s3-bucket main.tf
```