Allow TFREGISTRY_BASE_URL to set the host of the Terraform registry
In #130, I added opentofuRegistryModule and opentofuRegistryProvider to
the --source type in the release latest and list commands, but it turned
out that this was not a good design.

When considering OpenTofu support for the lock command, I noticed that
the lock file records the hostname of the registry. However, this cannot
be retrieved from the required_providers block. We could pass the
hostname as an argument, but GitHub and GitLab already allow the BaseURL
to be passed from an environment variable, so it would be better to
align it with this.

Also, opentofuRegistryModule and opentofuRegistryProvider only replace
BaseURL and call compatible APIs at the protocol level. They are
redundant and should be removed before cutting a new release.