Unify tfregistry config for release and lock packages
To support OpenTofu, we have made it possible to replace the BaseURL of
the Terraform Registry in the release package, but there are still
scattered places where BaseURL is handled.

We will move tfregistry-related implementations from the release and
lock packages to the tfregistry package to unify them, so that the
BaseURL can also be replaced in the lock package.

This is a breaking change as Go packages, but as a CLI, it should not
affect end users.