Add support for updating .terraform.lock.hcl using OpenTofu registry
To make the .terraform.lock.hcl update compatible with OpenTofu, the
hash value is different between the Terraform Registry and OpenTofu
Registry, so we need to replace the BaseURL, download the provider from
the OpenTofu Registry, and calculate the hash value. Also, since the
FQDN of the provider recorded in the lock contains the hostname, this
must be replaced as well.