Fix the build issue with the replace directive
Fixes #135

When I used a forked version of hashicorp/terraform-config-inspect to
support the .tofu extension in #128, I had forgotten that the replace
directive does not work when go install from outside the repository.

So far, I have not found a reliable OpenTofu equivalent of
terraform-config-inspect, and there is no possibility that the forked
version will be merged into upstream.

In general, changing the import path can cause conflicts with upstream,
but fortunately, it is not used in too many places. I'm waiting for the
official OpenTofu version to be released, but for now, I'll align the
import path with the forked version and remove the replace directive.

https://github.com/minamijoyo/terraform-config-inspect/commit/6ad6eb27d3c9437faaf5e0f793f3892701a90ada